CLI config flags Fixes #64
