+++ cities.json
@@ -751,7 +751,7 @@
     "longitude": 15.0444
   },
   {
-    "name": "Lom\u00e9",
+    "name": "Lome",
     "country": "Togo",
     "latitude": 6.13748,
     "longitude": 1.21227
@@ -781,7 +781,7 @@
     "longitude": 10.16579
   },
   {
-    "name": "\u0130stanbul",
+    "name": "Istanbul",
     "country": "Turkey",
     "latitude": 41.01384,
     "longitude": 28.94966
